Refactor github client
