+++ src/bioregistry/gh/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+"""GitHub utilities."""
